% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-arima_impl.R
\name{Arima_predict_impl}
\alias{Arima_predict_impl}
\title{Low-Level ARIMA function for translating modeltime to forecast}
\usage{
Arima_predict_impl(object, new_data, ...)
}
\arguments{
\item{object}{An object of class \code{model_fit}}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{...}{Additional arguments passed to \code{forecast::forcast.Arima()}}
}
\description{
Low-Level ARIMA function for translating modeltime to forecast
}
