% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arima_workflow_tuned}
\alias{arima_workflow_tuned}
\title{Example ARIMA Tuning Results}
\format{
An object of class \code{tune_results} (inherits from \code{time_series_cv}, \code{rset}, \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 4 columns.
}
\usage{
arima_workflow_tuned
}
\value{
This is the output of \code{tune_grid()} for an ARIMA model created
with \code{\link[=arima_reg]{arima_reg()}}.
}
\description{
These objects are the results of an analysis of the
M750 data set, which came from the M4 Forecast Competition.
}
\examples{
arima_workflow_tuned

}
\keyword{datasets}
