% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-adam.R
\name{Auto_adam_predict_impl}
\alias{Auto_adam_predict_impl}
\title{Bridge prediction function for AUTO ADAM models}
\usage{
Auto_adam_predict_impl(object, new_data, ...)
}
\arguments{
\item{object}{An object of class \code{model_fit}.}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{...}{Additional arguments passed to \code{smooth::auto.adam()}}
}
\description{
Bridge prediction function for AUTO ADAM models
}
\keyword{internal}
