% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-arima_boost.R
\name{auto_arima_xgboost_predict_impl}
\alias{auto_arima_xgboost_predict_impl}
\title{Bridge prediction Function for ARIMA-XGBoost Models}
\usage{
auto_arima_xgboost_predict_impl(object, new_data, ...)
}
\arguments{
\item{object}{An object of class \code{model_fit}}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{...}{Additional arguments passed to \code{predict.xgb.Booster()}}
}
\description{
Bridge prediction Function for ARIMA-XGBoost Models
}
