% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime-refit.R
\name{check_class_and_single}
\alias{check_class_and_single}
\title{These are not intended for use by the general public.}
\usage{
check_class_and_single(x, cls = "numeric")
}
\arguments{
\item{x}{An object}

\item{cls}{A character vector of possible classes}
}
\value{
Control information
}
\description{
These are not intended for use by the general public.
}
