% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime-refit.R
\name{check_class_integer}
\alias{check_class_integer}
\title{These are not intended for use by the general public.}
\usage{
check_class_integer(x)
}
\arguments{
\item{x}{A number}
}
\value{
A logical value
}
\description{
These are not intended for use by the general public.
}
