% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime-table.R
\name{combine_model_tables}
\alias{combine_model_tables}
\title{Combine multiple Modeltime Tables into a single Modeltime Table}
\usage{
combine_model_tables(...)
}
\arguments{
\item{...}{A Modeltime Tables (class \code{mdl_time_tbl})}
}
\description{
Combine multiple Modeltime Tables into a single Modeltime Table
}
\details{
This function can be used with both uncalibrated and calibrated
modeltime tables.
}
