% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime-refit.R
\name{control_refit}
\alias{control_refit}
\title{Control Refit}
\usage{
control_refit(verbose = FALSE, allow_par = TRUE, cores = 1, packages = NULL)
}
\arguments{
\item{verbose}{Logical to control printing.}

\item{allow_par}{Logical to allow parallel computation}

\item{cores}{Number of cores for computation}

\item{packages}{An optional character string of R package names that should be loaded
(by namespace) during parallel processing.}
}
\value{
A List with the information.
}
\description{
Control aspects of the \code{modeltime_refit()} process.
}
\seealso{
\code{\link[=modeltime_refit]{modeltime_refit()}}
}
