% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime-fit-workflowset.R
\name{control_workflowset}
\alias{control_workflowset}
\title{Control aspects of the modeltime_fit_workflowset process.}
\usage{
control_workflowset(
  allow_par = TRUE,
  cores = 1,
  packages = c("modeltime", "parsnip", "dplyr", "stats", "lubridate", "tidymodels",
    "timetk"),
  verbose = FALSE
)
}
\arguments{
\item{allow_par}{Logical to allow parallel computation}

\item{cores}{Number of cores for computation}

\item{packages}{Packages that the user wants to send to the workers during parallelization.}

\item{verbose}{Logical to control printing.}
}
\value{
A List with the information.
}
\description{
Control aspects of the modeltime_fit_workflowset process.
}
\seealso{
\code{\link[=modeltime_fit_workflowset]{modeltime_fit_workflowset()}}
}
