% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-exp_smoothing.R
\name{croston_fit_impl}
\alias{croston_fit_impl}
\title{Low-Level Exponential Smoothing function for translating modeltime to forecast}
\usage{
croston_fit_impl(x, y, alpha = 0.1, ...)
}
\arguments{
\item{x}{A dataframe of xreg (exogenous regressors)}

\item{y}{A numeric vector of values to fit}

\item{alpha}{Value of alpha. Default value is 0.1.}

\item{...}{Additional arguments passed to \code{forecast::ets}}
}
\description{
Low-Level Exponential Smoothing function for translating modeltime to forecast
}
\keyword{internal}
