% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials-croston_params.R
\name{croston_params}
\alias{croston_params}
\alias{alpha}
\title{Tuning Parameters for Croston Models}
\usage{
alpha(range = c(0, 1), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::log10_trans()} or \code{scales::reciprocal_trans()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
Tuning Parameters for Croston Models
}
\details{
The main parameters for Croston models are:
\itemize{
\item \code{alpha}: The order of the non-seasonal auto-regressive (AR) terms.
}
}
\examples{
alpha()


}
