% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime-recursive.R
\name{.prepare_transform}
\alias{.prepare_transform}
\alias{.prepare_panel_transform}
\title{Prepare Recursive Transformations}
\usage{
.prepare_transform(.transform)

.prepare_panel_transform(.transform)
}
\arguments{
\item{.transform}{A transformation function}
}
\value{
A function that applies a recursive transformation
}
\description{
Prepare Recursive Transformations
}
\keyword{internal}
