% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-exp_smoothing.R
\name{ets_predict_impl}
\alias{ets_predict_impl}
\title{Bridge prediction function for Exponential Smoothing models}
\usage{
ets_predict_impl(object, new_data, ...)
}
\arguments{
\item{object}{An object of class \code{model_fit}.}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{...}{Additional arguments passed to \code{forecast::ets()}}
}
\description{
Bridge prediction function for Exponential Smoothing models
}
\keyword{internal}
