% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-model_descriptions.R
\name{get_arima_description}
\alias{get_arima_description}
\title{Get model descriptions for Arima objects}
\source{
\itemize{
\item Forecast R Package, \code{forecast:::arima.string()}
}
}
\usage{
get_arima_description(object, padding = FALSE)
}
\arguments{
\item{object}{Objects of class \code{Arima}}

\item{padding}{Whether or not to include padding}
}
\description{
Get model descriptions for Arima objects
}
\examples{
library(forecast)

arima_fit <- forecast::Arima(1:10)

get_arima_description(arima_fit)


}
