% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-model_descriptions.R
\name{get_tbats_description}
\alias{get_tbats_description}
\title{Get model descriptions for TBATS objects}
\source{
\itemize{
\item Forecast R Package, \code{forecast:::as.character.tbats()}
}
}
\usage{
get_tbats_description(object)
}
\arguments{
\item{object}{Objects of class \code{tbats}}
}
\description{
Get model descriptions for TBATS objects
}
