% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-hierarchical_reg.R
\name{hierarchical_fit_impl}
\alias{hierarchical_fit_impl}
\title{Low-Level Temporaral Hierarchical function for translating modeltime to forecast}
\usage{
hierarchical_fit_impl(
  x,
  y,
  comb = c("struc", "mse", "ols", "bu", "shr", "sam"),
  usemodel = c("ets", "arima", "theta", "naive", "snaive"),
  ...
)
}
\arguments{
\item{x}{A dataframe of xreg (exogenous regressors)}

\item{y}{A numeric vector of values to fit}

\item{comb}{Combination method of temporal hierarchies}

\item{usemodel}{Model used for forecasting each aggregation level}

\item{...}{Additional arguments passed to \code{forecast::ets}}
}
\description{
Low-Level Temporaral Hierarchical function for translating modeltime to forecast
}
