% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-hierarchical_reg.R
\name{hierarchical_predict_impl}
\alias{hierarchical_predict_impl}
\title{Bridge prediction function for TEMPORAL HIERARCHICAL models}
\usage{
hierarchical_predict_impl(object, new_data, ...)
}
\arguments{
\item{object}{An object of class \code{model_fit}}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{...}{Additional arguments passed to \code{stats::predict()}}
}
\description{
Bridge prediction function for TEMPORAL HIERARCHICAL models
}
