% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-is_modeltime.R
\name{is_calibrated}
\alias{is_calibrated}
\title{Test if a Modeltime Table has been calibrated}
\usage{
is_calibrated(object)
}
\arguments{
\item{object}{An object to detect if is a Calibrated Modeltime Table}
}
\description{
This function returns \code{TRUE} for objects that contains columns
".type" and ".calibration_data"
}
\keyword{internal}
