% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-is_modeltime.R
\name{is_modeltime_model}
\alias{is_modeltime_model}
\title{Test if object contains a fitted modeltime model}
\usage{
is_modeltime_model(object)
}
\arguments{
\item{object}{An object to detect if contains a fitted modeltime model}
}
\description{
This function returns \code{TRUE} for trained workflows and parsnip objects
that contain modeltime models
}
\keyword{internal}
