% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-is_modeltime.R
\name{is_modeltime_table}
\alias{is_modeltime_table}
\title{Test if object is a Modeltime Table}
\usage{
is_modeltime_table(object)
}
\arguments{
\item{object}{An object to detect if is a Modeltime Table}
}
\description{
This function returns \code{TRUE} for objects that contain class \code{mdl_time_tbl}
}
\keyword{internal}
