% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-is_modeltime.R
\name{is_residuals}
\alias{is_residuals}
\title{Test if a table contains residuals.}
\usage{
is_residuals(object)
}
\arguments{
\item{object}{An object to detect if it provides from modeltime::modeltime_residuals().}
}
\description{
This function returns \code{TRUE} for objects that contains the column name '.residuals'.
}
\keyword{internal}
