% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-control-par.R
\name{load_namespace}
\alias{load_namespace}
\title{These are not intended for use by the general public.}
\usage{
load_namespace(x, full_load)
}
\arguments{
\item{x}{A vector}

\item{full_load}{A vector}
}
\value{
Control information
}
\description{
These are not intended for use by the general public.
}
\keyword{internal}
