% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_modeltime_fit.R
\name{log_extractors}
\alias{log_extractors}
\alias{extract_nested_test_accuracy}
\alias{extract_nested_test_forecast}
\alias{extract_nested_error_report}
\alias{extract_nested_best_model_report}
\alias{extract_nested_future_forecast}
\title{Log Extractor Functions for Modeltime Nested Tables}
\usage{
extract_nested_test_accuracy(object)

extract_nested_test_forecast(object)

extract_nested_error_report(object)

extract_nested_best_model_report(object)

extract_nested_future_forecast(object)
}
\arguments{
\item{object}{A nested modeltime table}
}
\description{
Extract logged information calculated during the \code{modeltime_nested_fit()},
\code{modeltime_nested_select_best()}, and \code{modeltime_nested_refit()} processes.
}
