% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m750}
\alias{m750}
\title{The 750th Monthly Time Series used in the M4 Competition}
\format{
A \code{tibble} with 306 rows and 3 variables:
\itemize{
\item \code{id} Factor. Unique series identifier
\item \code{date} Date. Timestamp information. Monthly format.
\item \code{value} Numeric. Value at the corresponding timestamp.
}
}
\source{
\itemize{
\item M4 Competition Website: https://www.unic.ac.cy/iff/research/forecasting/m-competitions/m4/
}
}
\usage{
m750
}
\description{
The 750th Monthly Time Series used in the M4 Competition
}
\examples{
m750

}
\keyword{datasets}
