% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m750_models}
\alias{m750_models}
\title{Three (3) Models trained on the M750 Data (Training Set)}
\format{
An \code{time_series_cv} object with 6 slices of Time Series Cross Validation
resamples made on the \code{training(m750_splits)}
}
\usage{
m750_models
}
\description{
Three (3) Models trained on the M750 Data (Training Set)
}
\details{
\if{html}{\out{<div class="r">}}\preformatted{library(modeltime)
m750_models <- modeltime_table(
    wflw_fit_arima,
    wflw_fit_prophet,
    wflw_fit_glmnet
)
}\if{html}{\out{</div>}}
}
\examples{
library(modeltime)

m750_models



}
\keyword{datasets}
