% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m750_splits}
\alias{m750_splits}
\title{The results of train/test splitting the M750 Data}
\format{
An \code{rsplit} object split into approximately 23.5-years of training data
and 2-years of testing data
}
\usage{
m750_splits
}
\description{
The results of train/test splitting the M750 Data
}
\details{
\if{html}{\out{<div class="r">}}\preformatted{library(timetk)
m750_splits <- time_series_split(m750, assess = "2 years", cumulative = TRUE)
}\if{html}{\out{</div>}}
}
\examples{
library(rsample)

m750_splits

training(m750_splits)

}
\keyword{datasets}
