% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m750_training_resamples}
\alias{m750_training_resamples}
\title{The Time Series Cross Validation Resamples the M750 Data (Training Set)}
\format{
An \code{time_series_cv} object with 6 slices of Time Series Cross Validation
resamples made on the \code{training(m750_splits)}
}
\usage{
m750_training_resamples
}
\description{
The Time Series Cross Validation Resamples the M750 Data (Training Set)
}
\details{
\if{html}{\out{<div class="r">}}\preformatted{library(timetk)
m750_training_resamples <- time_series_cv(
    data        = training(m750_splits),
    assess      = "2 years",
    skip        = "2 years",
    cumulative  = TRUE,
    slice_limit = 6
)
}\if{html}{\out{</div>}}
}
\examples{
library(rsample)

m750_training_resamples



}
\keyword{datasets}
