% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yardstick-metric-sets.R
\name{maape.data.frame}
\alias{maape.data.frame}
\title{Mean Arctangent Absolute Percentage Error}
\usage{
\method{maape}{data.frame}(data, truth, estimate, na_rm = TRUE, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the truth and estimate columns.}

\item{truth}{The column identifier for the true results (that is numeric).}

\item{estimate}{The column identifier for the predicted results (that is also numeric).}

\item{na_rm}{Not in use...NA values managed by TSrepr::maape}

\item{...}{Not currently in use}
}
\description{
This is basically a wrapper to the function of \code{TSrepr::maape()}.
}
