% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yardstick-metric-sets.R
\name{maape_vec}
\alias{maape_vec}
\title{Mean Arctangent Absolute Percentage Error}
\usage{
maape_vec(truth, estimate, na_rm = TRUE, ...)
}
\arguments{
\item{truth}{Numeric vector of ground-truth values.}

\item{estimate}{Numeric vector of predicted values.}

\item{na_rm}{Remove missing values before computation.}

\item{...}{Unused.}
}
\description{
Mean Arctangent Absolute Percentage Error
}
