% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested-modeltime_data_prep.R
\name{make_ts_splits}
\alias{make_ts_splits}
\title{Generate a Time Series Train/Test Split Indicies}
\usage{
make_ts_splits(.data, .length_test, .length_train = NULL)
}
\arguments{
\item{.data}{A data frame containing ordered time seried data (ascending)}

\item{.length_test}{The number of rows to include in the test set}

\item{.length_train}{Optional. The number of rows to include in the training set.
If NULL, returns all remaining row indicies.}
}
\value{
A list containing train_idx and test_idx
}
\description{
Makes fast train/test split indicies for time series.
}
\keyword{internal}
