% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime-forecast.R
\name{mdl_time_forecast.recursive}
\alias{mdl_time_forecast.recursive}
\title{For now, it's considered to be applied for ensemble objects}
\usage{
\method{mdl_time_forecast}{recursive}(
  object,
  calibration_data,
  new_data = NULL,
  h = NULL,
  actual_data = NULL,
  bind_actual = TRUE,
  keep_data = FALSE,
  arrange_index = FALSE,
  ...
)
}
\description{
For now, it's considered to be applied for ensemble objects
}
