% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime-fit-workflowset.R
\name{modeltime_fit_workflowset}
\alias{modeltime_fit_workflowset}
\title{Fit a \code{workflowset} object to one or multiple time series}
\usage{
modeltime_fit_workflowset(workflowset, series, control = control_workflowset())
}
\arguments{
\item{workflowset}{A workflow_set object, generated with the workflowsets::workflow_set function.}

\item{series}{A \code{tibble} that contains data to fit the models.}

\item{control}{An object used to modify the fitting process.}
}
\value{
A Modeltime Table containing one or more fitted models.
}
\description{
This is a wrapper for \code{fit()} that takes a
\code{workflowset} object and fits each model on one or multiple
time series either sequentially or in parallel.
}
\seealso{
\code{\link[=control_workflowset]{control_workflowset()}}
}
