% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested-modeltime-forecast.R
\name{modeltime_nested_forecast}
\alias{modeltime_nested_forecast}
\title{Modeltime Nested Forecast}
\usage{
modeltime_nested_forecast(
  object,
  h = NULL,
  include_actual = TRUE,
  conf_interval = 0.95,
  control = control_nested_forecast()
)
}
\description{
Make a new forecast from a Nested Modeltime Table.
}
