% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_modeltime_fit.R
\name{modeltime_nested_tables}
\alias{modeltime_nested_tables}
\alias{modeltime_nested_accuracy}
\alias{modeltime_nested_test_forecast}
\alias{modeltime_nested_error_report}
\alias{modeltime_nested_future_forecast}
\alias{modeltime_nested_best_model_report}
\title{Logging Functions for Modeltime Nested Tables}
\usage{
modeltime_nested_accuracy(object)

modeltime_nested_test_forecast(object)

modeltime_nested_error_report(object)

modeltime_nested_future_forecast(object)

modeltime_nested_best_model_report(object)
}
\arguments{
\item{object}{A nested modeltime table}
}
\description{
Logging Functions for Modeltime Nested Tables
}
