% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime-refit.R
\name{modeltime_refit}
\alias{modeltime_refit}
\alias{modeltime_refit_xy}
\title{Refit one or more trained models to new data}
\usage{
modeltime_refit(object, data, control = NULL, ...)

modeltime_refit_xy(object, x, y, control = NULL, ...)
}
\arguments{
\item{object}{A fitted model object that is either (1) a workflow that has been fit by \code{\link[=fit.workflow]{fit.workflow()}} or
(2) a parsnip model that has been fit using \code{\link[=fit.model_spec]{fit.model_spec()}}}

\item{data}{A \code{tibble}}

\item{control}{Either \code{\link[=control_parsnip]{control_parsnip()}} or \code{\link[=control_workflow]{control_workflow()}} depending on the object.
If NULL, created automatically.}

\item{...}{Additional arguments passed to \code{\link[=fit]{fit()}}.}

\item{x}{A matrix or data frame of predictors.}

\item{y}{A vector, matrix or data frame of outcome data.}
}
\value{
Depending on the object provided, either:
\itemize{
\item Single Parsnip Model: A fitted parsnip model (\code{model_fit}) is returned.
\item Single Workflow: A trained \code{workflow} is returned.
\item A Modeltime Table containing one or more trained models: A trained \code{mdl_time_tbl} table
with one or more models that have been retrained is returned.
}
}
\description{
This is a wrapper for \code{fit()} and \code{fit_xy()} that takes a
trained parsnip model or workflow and retrains on \emph{new data} using the parameters
and preprocessing (formulas or recipes) used during the training process:
\itemize{
\item Automated models (e.g. "auto arima") will have parameters recalculated.
\item Non-automated models (e.g. "arima") will have parameters preserved.
\item All preprocessing steps will be reused on the data
}
}
\details{
TODO
}
\examples{
# TODO

}
