% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime_residuals_tests.R
\name{modeltime_residuals_tests}
\alias{modeltime_residuals_tests}
\title{Extract Residuals Tests}
\usage{
modeltime_residuals_tests(
  object,
  new_data = NULL,
  quiet = TRUE,
  box_pierce_lag = 1,
  box_pierce_fitdf = 0,
  ljung_box_lag = 1,
  ljung_box_fitdf = 0,
  ...
)
}
\arguments{
\item{object}{A \code{tibble} extracted from modeltime::modeltime_residuals().}

\item{new_data}{A \code{tibble} to predict and calculate residuals on.
If provided, overrides any calibration data.}

\item{quiet}{Hide errors (\code{TRUE}, the default), or display them as they occur?}

\item{box_pierce_lag}{the statistic will be based on lag autocorrelation coefficients. Default: 1}

\item{box_pierce_fitdf}{number of degrees of freedom to be subtracted. Default: 0}

\item{ljung_box_lag}{the statistic will be based on lag autocorrelation coefficients. Default: 1}

\item{ljung_box_fitdf}{number of degrees of freedom to be subtracted. Default: 0}

\item{...}{Not currently used}
}
\value{
A tibble with with the p-values of the calculated statistical tests.
}
\description{
This is a convenience function to calculate some statistical tests on the residuals models. Currently, the
following statistics are calculated: the shapiro.test to check the normality of the residuals, the box-pierce
and ljung-box tests and the durbin watson test to check the autocorrelation of the residuals. In all cases
the p-values are returned.
}
\details{
For more information about the parameters associated with the Box Pierce and Ljung Box tests check
?Box.Test
}
\examples{
library(tidyverse)
library(lubridate)
library(timetk)
library(parsnip)
library(rsample)

# Data
m750 <- m4_monthly \%>\% filter(id == "M750")

# Split Data 80/20
splits <- initial_time_split(m750, prop = 0.9)

# --- MODELS ---

# Model 1: auto_arima ----
model_fit_arima <- arima_reg() \%>\%
    set_engine(engine = "auto_arima") \%>\%
    fit(value ~ date, data = training(splits))


# ---- MODELTIME TABLE ----

models_tbl <- modeltime_table(
    model_fit_arima
)

# ---- RESIDUALS ----

# In-Sample
models_tbl \%>\%
    modeltime_calibrate(new_data = training(splits)) \%>\%
    modeltime_residuals() \%>\%
    modeltime_residuals_tests()

# Out-of-Sample
models_tbl \%>\%
    modeltime_calibrate(new_data = testing(splits)) \%>\%
    modeltime_residuals() \%>\%
    modeltime_residuals_tests()


}
