% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime-table.R
\name{modeltime_table}
\alias{modeltime_table}
\title{Scale forecast analysis with a Modeltime Table}
\usage{
modeltime_table(...)
}
\arguments{
\item{...}{Fitted model or workflow objects}
}
\description{
Designed to perform forecasts at scale using models created with
\code{modeltime}, \code{parsnip}, \code{workflows}, and regression modeling extensions
in the \code{tidymodels} ecosystem.
}
\examples{
library(tidyverse)
library(lubridate)
library(timetk)
library(workflows)
library(parsnip)
library(recipes)
library(rsample)

# Data
m750 <- m4_monthly \%>\% filter(id == "M750")

# Split Data 80/20
splits <- initial_time_split(m750, prop = 0.9)

# --- MODELS ---

# Model 1: auto_arima ----
model_fit_no_boost <- arima_reg() \%>\%
    set_engine(engine = "auto_arima") \%>\%
    fit(value ~ date, data = training(splits))

# Model 2: arima_boost ----
model_fit_boosted <- arima_boost(
    min_n = 2,
    learn_rate = 0.015
) \%>\%
    set_engine(engine = "auto_arima_xgboost") \%>\%
    fit(value ~ date + as.numeric(date) + month(date, label = TRUE),
        data = training(splits))

# ---- COMPARE ----

models_tbl <- modeltime_table(
    model_fit_no_boost,
    model_fit_boosted
)

models_tbl \%>\%
    modeltime_accuracy(new_data = testing(splits))

}
