% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-naive_reg.R
\name{naive_fit_impl}
\alias{naive_fit_impl}
\title{Low-Level NAIVE Forecast}
\usage{
naive_fit_impl(x, y, id = NULL, seasonal_period = "auto", ...)
}
\arguments{
\item{x}{A dataframe of xreg (exogenous regressors)}

\item{y}{A numeric vector of values to fit}

\item{id}{An optional ID feature to identify different time series. Should be a quoted name.}

\item{seasonal_period}{Not used for NAIVE forecast but here for consistency with SNAIVE}

\item{...}{Not currently used}
}
\description{
Low-Level NAIVE Forecast
}
\keyword{internal}
