% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime-recursive.R
\name{panel_tail}
\alias{panel_tail}
\title{Filter the last N rows (Tail) for multiple time series}
\usage{
panel_tail(data, id, n)
}
\arguments{
\item{data}{A data frame}

\item{id}{An "id" feature indicating which column differentiates the time series panels}

\item{n}{The number of rows to filter}
}
\value{
A data frame
}
\description{
Filter the last N rows (Tail) for multiple time series
}
\examples{
library(timetk)

# Get the last 6 observations from each group
m4_monthly \%>\%
    panel_tail(id = id, n = 6)

}
\seealso{
\itemize{
\item \code{\link[=recursive]{recursive()}} - used to generate recursive autoregressive models
}
}
