% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime-forecast-plot.R
\name{plot_modeltime_forecast}
\alias{plot_modeltime_forecast}
\title{Interactive Plotting for One or More Time Series}
\usage{
plot_modeltime_forecast(
  .data,
  .title = "Forecast Plot",
  .x_lab = "",
  .y_lab = "",
  .color_lab = "Legend",
  .interactive = TRUE,
  ...
)
}
\arguments{
\item{.data}{A \code{tibble} or \code{data.frame} with '.id', '.index', and .value' columns}

\item{.title}{Title for the plot}

\item{.x_lab}{X-axis label for the plot}

\item{.y_lab}{Y-axis label for the plot}

\item{.color_lab}{Legend label if a \code{color_var} is used.}

\item{.interactive}{Returns either a static (\code{ggplot2}) visualization or an interactive (\code{plotly}) visualization}

\item{...}{Additional arguments passed to \code{\link[timetk:plot_time_series]{timetk::plot_time_series()}}.}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot containing a forecast
}
\description{
A workhorse time-series plotting function that generates interactive \code{plotly} plots,
consolidates 20+ lines of \code{ggplot2} code, and scales well to many time series.
}
\details{
TODO
}
\examples{
library(dplyr)
library(parsnip)
library(rsample)
library(timetk)
library(modeltime)

# Data
m750 <- m4_monthly \%>\% filter(id == "M750")
m750

# Split Data 80/20
splits <- initial_time_split(m750, prop = 0.8)

# Model Spec
model_spec <- arima_reg() \%>\%
    set_engine("forecast")

# Fit Spec
model_fit <- model_spec \%>\%
    fit(log(value) ~ date, data = training(splits))

# --- VISUALIZE FORECAST ---

# Combining forecast with actual values
model_fit \%>\%
    modeltime_forecast(h = "3 years", actual_data = training(splits)) \%>\%
    plot_modeltime_forecast(.interactive = FALSE)

}
