% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-prophet_reg.R
\name{prophet_predict_impl}
\alias{prophet_predict_impl}
\title{Bridge prediction function for PROPHET models}
\usage{
prophet_predict_impl(object, new_data, ...)
}
\arguments{
\item{object}{An object of class \code{model_fit}.}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{...}{Additional arguments passed to \code{prophet::predict()}}
}
\description{
Bridge prediction function for PROPHET models
}
\keyword{internal}
