% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-modeltime_residuals.R
\name{pull_modeltime_residuals}
\alias{pull_modeltime_residuals}
\title{Extracts modeltime residuals data from a Modeltime Model}
\usage{
pull_modeltime_residuals(object)
}
\arguments{
\item{object}{A fitted \code{parsnip} / \code{modeltime} model or \code{workflow}}
}
\value{
A \code{tibble} containing the model timestamp, actual, fitted, and residuals data
}
\description{
If a modeltime model contains \code{data} with residuals information,
this function will extract the data frame.
}
