% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parsnip-helpers.R
\name{pull_parsnip_preprocessor}
\alias{pull_parsnip_preprocessor}
\title{Pulls the Formula from a Fitted Parsnip Model Object}
\usage{
pull_parsnip_preprocessor(object)
}
\arguments{
\item{object}{A fitted parsnip model \code{model_fit} object}
}
\value{
A formula using \code{stats::formula()}
}
\description{
Pulls the Formula from a Fitted Parsnip Model Object
}
