% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-naive_reg.R
\name{snaive_fit_impl}
\alias{snaive_fit_impl}
\title{Low-Level SNAIVE Forecast}
\usage{
snaive_fit_impl(x, y, id = NULL, seasonal_period = "auto", ...)
}
\arguments{
\item{x}{A dataframe of xreg (exogenous regressors)}

\item{y}{A numeric vector of values to fit}

\item{id}{An optional ID feature to identify different time series. Should be a quoted name.}

\item{seasonal_period}{The seasonal period to forecast into the future}

\item{...}{Not currently used}
}
\description{
Low-Level SNAIVE Forecast
}
\keyword{internal}
