% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-naive_reg.R
\name{snaive_predict_impl}
\alias{snaive_predict_impl}
\title{Bridge prediction function for SNAIVE Models}
\usage{
snaive_predict_impl(object, new_data)
}
\arguments{
\item{object}{An object of class \code{model_fit}.}

\item{new_data}{A rectangular data object, such as a data frame.}
}
\description{
Bridge prediction function for SNAIVE Models
}
\keyword{internal}
