% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-seasonal_reg.R
\name{stlm_ets_fit_impl}
\alias{stlm_ets_fit_impl}
\title{Low-Level stlm function for translating modeltime to forecast}
\usage{
stlm_ets_fit_impl(
  x,
  y,
  period_1 = "auto",
  period_2 = NULL,
  period_3 = NULL,
  ...
)
}
\arguments{
\item{x}{A dataframe of xreg (exogenous regressors)}

\item{y}{A numeric vector of values to fit}

\item{period_1}{(required) First seasonal frequency. Uses "auto" by default. A character phrase
of "auto" or time-based phrase of "2 weeks" can be used if a date or date-time variable is provided.}

\item{period_2}{(optional) First seasonal frequency. Uses "auto" by default. A character phrase
of "auto" or time-based phrase of "2 weeks" can be used if a date or date-time variable is provided.}

\item{period_3}{(optional) First seasonal frequency. Uses "auto" by default. A character phrase
of "auto" or time-based phrase of "2 weeks" can be used if a date or date-time variable is provided.}

\item{...}{Additional arguments passed to \code{forecast::stlm()}}
}
\description{
Low-Level stlm function for translating modeltime to forecast
}
\keyword{internal}
