% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-temporal_hierarchy.R
\name{temporal_hier_predict_impl}
\alias{temporal_hier_predict_impl}
\title{Bridge prediction function for TEMPORAL HIERARCHICAL models}
\usage{
temporal_hier_predict_impl(object, new_data, ...)
}
\arguments{
\item{object}{An object of class \code{model_fit}.}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{...}{Additional arguments passed to \code{stats::predict()}}
}
\description{
Bridge prediction function for TEMPORAL HIERARCHICAL models
}
\keyword{internal}
