% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials-temporal_hierarchy_params.R
\name{temporal_hierarchy_params}
\alias{temporal_hierarchy_params}
\alias{combination_method}
\alias{use_model}
\title{Tuning Parameters for TEMPORAL HIERARCHICAL Models}
\usage{
combination_method(values = c("struc", "mse", "ols", "bu", "shr", "sam"))

use_model()
}
\arguments{
\item{values}{A character string of possible values.}
}
\description{
Tuning Parameters for TEMPORAL HIERARCHICAL Models
}
\details{
The main parameters for Temporal Hierarchical models are:
\itemize{
\item \code{combination_method}: Combination method of temporal hierarchies.
\item \code{use_model}: Model used for forecasting each aggregation level.
}
}
\examples{
combination_method()

use_model()

}
