% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-exp_smoothing.R
\name{theta_fit_impl}
\alias{theta_fit_impl}
\title{Low-Level Exponential Smoothing function for translating modeltime to forecast}
\usage{
theta_fit_impl(x, y, ...)
}
\arguments{
\item{x}{A dataframe of xreg (exogenous regressors)}

\item{y}{A numeric vector of values to fit}

\item{...}{Additional arguments passed to \code{forecast::ets}}
}
\description{
Low-Level Exponential Smoothing function for translating modeltime to forecast
}
\keyword{internal}
