% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials-ts_params.R
\name{time_series_params}
\alias{time_series_params}
\alias{period}
\title{Tuning Parameters for Time Series (ts-class) Models}
\usage{
period(range = c(1L, 12L), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::log10_trans()} or \code{scales::reciprocal_trans()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
Tuning Parameters for Time Series (ts-class) Models
}
\details{
Time series models (e.g. \code{Arima()} and \code{ets()}) use \code{\link[stats:ts]{stats::ts()}}
to apply seasonality. We can do the same process using the following
general time series parameter:
\itemize{
\item \code{period}: The periodic nature of the seasonality. Set 1 for non-seasonal.
}
}
\examples{
period()


}
