% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble-type_sum.R
\name{type_sum.mdl_time_tbl}
\alias{type_sum.mdl_time_tbl}
\title{Succinct summary of Modeltime Tables}
\usage{
\method{type_sum}{mdl_time_tbl}(x)
}
\arguments{
\item{x}{A \code{mdl_time_tbl} object to summarise.}
}
\value{
A character value.
}
\description{
\code{type_sum} controls how objects are shown when inside tibble
columns.
}
\keyword{internal}
