% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime-table.R
\name{update_model_description}
\alias{update_model_description}
\title{Update the model description by model id in a Modeltime Table}
\usage{
update_model_description(object, .model_id, .new_model_desc)
}
\arguments{
\item{object}{A Modeltime Table}

\item{.model_id}{A numeric value matching the .model_id that you want to update}

\item{.new_model_desc}{Text describing the new model description}
}
\description{
Update the model description by model id in a Modeltime Table
}
