% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime-refit.R
\name{val_class_and_single}
\alias{val_class_and_single}
\title{These are not intended for use by the general public.}
\usage{
val_class_and_single(x, cls = "numeric", where = NULL)
}
\arguments{
\item{x}{An object}

\item{cls}{A character vector of possible classes}

\item{where}{A character string for the calling function}
}
\value{
Control information
}
\description{
These are not intended for use by the general public.
}
