% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-window_reg.R
\name{window_fit_impl}
\alias{window_fit_impl}
\title{Low-Level Window Forecast}
\usage{
window_fit_impl(
  x,
  y,
  id = NULL,
  window_function = NULL,
  window_size = "all",
  ...
)
}
\arguments{
\item{x}{A dataframe of xreg (exogenous regressors)}

\item{y}{A numeric vector of values to fit}

\item{id}{An optional ID feature to identify different time series. Should be a quoted name.}

\item{window_function}{A function to apply to the window}

\item{window_size}{The period to apply the window function to}

\item{...}{Not currently used}
}
\description{
Low-Level Window Forecast
}
