% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-window_reg.R
\name{window_function_fit_impl}
\alias{window_function_fit_impl}
\title{Low-Level Window Forecast}
\usage{
window_function_fit_impl(
  x,
  y,
  id = NULL,
  window_size = "all",
  window_function = NULL,
  ...
)
}
\arguments{
\item{x}{A dataframe of xreg (exogenous regressors)}

\item{y}{A numeric vector of values to fit}

\item{id}{An optional ID feature to identify different time series. Should be a quoted name.}

\item{window_size}{The period to apply the window function to}

\item{window_function}{A function to apply to the window. The default is \code{mean()}.}

\item{...}{Additional arguments for the \code{window_function}. For example, it's
common to pass \code{na.rm = TRUE} for the mean forecast.}
}
\description{
Low-Level Window Forecast
}
\keyword{internal}
