% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-window_reg.R
\name{window_function_predict_impl}
\alias{window_function_predict_impl}
\title{Bridge prediction function for window Models}
\usage{
window_function_predict_impl(object, new_data)
}
\arguments{
\item{object}{An object of class \code{model_fit}.}

\item{new_data}{A rectangular data object, such as a data frame.}
}
\description{
Bridge prediction function for window Models
}
\keyword{internal}
