% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-window_reg.R
\name{window_lm_fit_impl}
\alias{window_lm_fit_impl}
\title{Low-Level Window Linear Regression Function}
\usage{
window_lm_fit_impl(x, y, id = NULL, window_size = "all", ...)
}
\arguments{
\item{x}{A dataframe of xreg (exogenous regressors)}

\item{y}{A numeric vector of values to fit}

\item{id}{An optional ID feature to identify different time series. Should be a quoted name.}

\item{window_size}{The period to apply the window function to}

\item{...}{Additional arguments for the \code{stats::lm()} function.}
}
\description{
Low-Level Window Linear Regression Function
}
