% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-xgboost.R
\name{xgboost_predict}
\alias{xgboost_predict}
\title{Wrapper for xgboost::predict}
\usage{
xgboost_predict(object, newdata, ...)
}
\arguments{
\item{object}{a model object for which prediction is desired.}

\item{newdata}{New data to be predicted}

\item{...}{additional arguments affecting the predictions produced.}
}
\description{
Wrapper for xgboost::predict
}
\keyword{internal}
