% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials-arima_params.R
\name{arima_params}
\alias{arima_params}
\alias{period}
\alias{p}
\alias{d}
\alias{q}
\alias{P}
\alias{D}
\alias{Q}
\title{Tuning Parameters for ARIMA Models}
\usage{
period(range = c(1L, 12L), trans = NULL)

p(range = c(0L, 5L), trans = NULL)

d(range = c(0L, 2L), trans = NULL)

q(range = c(0L, 5L), trans = NULL)

P(range = c(0L, 2L), trans = NULL)

D(range = c(0L, 1L), trans = NULL)

Q(range = c(0L, 2L), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::log10_trans()} or \code{scales::reciprocal_trans()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
Tuning Parameters for ARIMA Models
}
\details{
The main parameters for ARIMA models are:
\itemize{
\item \code{period}: The periodic nature of the seasonality. Set 1 for non-seasonal.
\item \code{p}: The order of the non-seasonal auto-regressive (AR) terms.
\item \code{d}: The order of integration for non-seasonal differencing.
\item \code{q}: The order of the non-seasonal moving average (MA) terms.
\item \code{P}: The order of the seasonal auto-regressive (SAR) terms.
\item \code{D}: The order of integration for seasonal differencing.
\item \code{Q}: The order of the seasonal moving average (SMA) terms.
}
}
\examples{
p()

d()

q()


}
