% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yardstick-metric-sets.R
\name{maape}
\alias{maape}
\title{Mean Arctangent Absolute Percentage Error}
\usage{
maape(data, ...)
}
\arguments{
\item{data}{A data frame.}
}
\description{
Mean Arctangent Absolute Percentage Error
}
