% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yardstick-metric-sets.R
\name{maape}
\alias{maape}
\title{Mean Arctangent Absolute Percentage Error}
\usage{
maape(data, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the truth and estimate columns.}

\item{...}{Not currently in use.}
}
\description{
This is basically a wrapper to the function of \code{TSrepr::maape()}.
}
