% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yardstick-metric-sets.R
\name{maape_vec}
\alias{maape_vec}
\title{Mean Arctangent Absolute Percentage Error}
\usage{
maape_vec(truth, estimate, na_rm = TRUE, ...)
}
\arguments{
\item{truth}{The column identifier for the true results (that is numeric).}

\item{estimate}{The column identifier for the predicted results (that is also numeric).}

\item{na_rm}{Not in use... \code{NA} values managed by \code{TSrepr::maape()}}

\item{...}{Not currently in use}
}
\description{
This is basically a wrapper to the function of \code{TSrepr::maape()}.
}
