% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime-forecast.R
\name{mdl_time_forecast}
\alias{mdl_time_forecast}
\title{Modeltime Forecast Helpers}
\usage{
mdl_time_forecast(
  object,
  new_data = NULL,
  h = NULL,
  conf_interval = 0.8,
  actual_data = NULL,
  ...
)
}
\arguments{
\item{object}{A Modeltime Table that has been calibrated with \code{\link[=modeltime_calibrate]{modeltime_calibrate()}}}

\item{new_data}{A \code{tibble} containing future information to forecast.}

\item{h}{The forecast horizon (can be used instead of \code{new_data} for
time series with no exogenous regressors). Always extends the calibration data.}

\item{conf_interval}{An estimated confidence interval based on the in-sample residuals}

\item{actual_data}{Reference data that is combined with the output tibble and given a \code{.key = "actual"}}

\item{...}{Additional arguments passed to \code{\link[=future_frame]{future_frame()}} for use with the \code{h} forecast horizon}
}
\value{
A tibble with forecast features
}
\description{
Used for low-level forecasting of modeltime, parnsip and workflow models.
These functions are not intended for user use.
}
\keyword{internal}
