% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime-refit.R
\name{mdl_time_refit}
\alias{mdl_time_refit}
\title{Modeltime Refit Helpers}
\usage{
mdl_time_refit(object, data, ..., control = NULL)
}
\arguments{
\item{object}{A Modeltime Table}

\item{data}{A \code{tibble} that contains data to retrain the model(s) using.}

\item{...}{Under construction. Additional arguments to control refitting.}

\item{control}{Under construction. Will be used to control refitting.}
}
\value{
A tibble with forecast features
}
\description{
Used for low-level refitting of modeltime, parnsip and workflow models
These functions are not intended for user use.
}
\keyword{internal}
