% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime-refit.R
\name{mdl_time_refit}
\alias{mdl_time_refit}
\title{Modeltime Refit Helpers}
\usage{
mdl_time_refit(object, data, ..., control = NULL)
}
\arguments{
\item{object}{A Modeltime Table}

\item{data}{A \code{tibble} that contains data to retrain the model(s) using.}

\item{...}{Additional arguments to control refitting.

\strong{Ensemble Model Spec (\code{modeltime.ensemble}):}

When making a meta-learner with \code{modeltime.ensemble::ensemble_model_spec()},
used to pass \code{resamples} argument containing results
from \code{modeltime.resample::modeltime_fit_resamples()}.}

\item{control}{Used to control verbosity and parallel processing.
See \code{\link[=control_refit]{control_refit()}}.}
}
\value{
A tibble with forecast features
}
\description{
Used for low-level refitting of modeltime, parnsip and workflow models
These functions are not intended for user use.
}
\keyword{internal}
