% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime_forecast.R
\name{modeltime_forecast}
\alias{modeltime_forecast}
\title{Forecast future data}
\usage{
modeltime_forecast(object, new_data = NULL, h = NULL, actual_data = NULL, ...)
}
\arguments{
\item{object}{A fitted model object that is either (1) a workflow that has been fit by \code{\link[=fit.workfow]{fit.workfow()}} or
(2) a parsnip model that has been fit using \code{\link[=fit.model_spec]{fit.model_spec()}}}

\item{new_data}{A \code{tibble} containing future information to forecast}

\item{h}{The forecast horizon (can be used instead of \code{new_data} for
time series with no exogenous regressors).}

\item{actual_data}{Data is combined with the output tibble and given an \code{.id = "actual"}}

\item{...}{Additional arguments passed to \code{\link[=future_frame]{future_frame()}} for use with the \code{h} forecast horizon}
}
\value{
A tibble with predictions and time-stamp data.
}
\description{
This is a wrapper for \code{predict()} that is useful for forecasting
future data from a fitted \code{workflow} (trained workflows object) or \code{model_fit} (trained parsnip model).
}
\details{
TODO
}
\examples{

# TODO



}
