% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control-par.R
\name{parallel_start}
\alias{parallel_start}
\alias{parallel_stop}
\title{Start parallel clusters using \code{parallel} package}
\usage{
parallel_start(...)

parallel_stop()
}
\arguments{
\item{...}{Parameters passed to \code{parallel::makeCluster()}}
}
\description{
Start parallel clusters using \code{parallel} package
}
\details{
Performs 3 Steps:
\enumerate{
\item Makes clusters using \code{parallel::makeCluster()}
\item Registers clusters using \code{doParallel::registerDoParallel()}
\item Adds \code{.libPaths()} using \code{parallel::clusterCall()}
}
}
\examples{

# Starts 2 clusters
parallel_start(2)

# Returns to sequential processing
parallel_stop()



}
