% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-modeltime_table.R
\name{pluck_modeltime_model}
\alias{pluck_modeltime_model}
\title{Extract model by model id in a Modeltime Table}
\usage{
pluck_modeltime_model(object, .model_id)
}
\arguments{
\item{object}{A Modeltime Table}

\item{.model_id}{A numeric value matching the .model_id that you want to update}
}
\description{
Extract model by model id in a Modeltime Table
}
